feat: Handled swapchain recreation

+++ src/vulkan_renderer/render_targets.rs
@@ -8,6 +8,8 @@ use super::vulkan_context::{SwapchainBuilder, VulkanContext};
 use builder::RenderTargetsBuilder;
 
 pub struct RenderTargets {
+    is_destroyed: bool,
+
     swapchain_device: ash::khr::swapchain::Device,
     swapchain: vk::SwapchainKHR,
     swapchain_images: Vec<vk::Image>,
@@ -37,30 +39,59 @@ impl RenderTargets {
     }
 
     pub fn render_pass(&self) -> vk::RenderPass {
+        #[cfg(feature = "validation_layers")] {
+            assert!(!self.is_destroyed,
+                    "RenderTargets::render_pass() was called after render_targets destruction");
+        }
         self.render_pass
     }
 
     pub fn framebuffers(&self) -> &[vk::Framebuffer] {
+        #[cfg(feature = "validation_layers")] {
+            assert!(!self.is_destroyed,
+                    "RenderTargets::framebuffers() was called after render_targets destruction");
+        }
         &self.framebuffers
     }
 
     pub fn swapchain_extent(&self) -> vk::Extent2D {
+        #[cfg(feature = "validation_layers")] {
+            assert!(!self.is_destroyed,
+                    "RenderTargets::swapchain_extent() was called after render_targets destruction");
+        }
         self.swapchain_extent
     }
 
     pub fn pipeline(&self) -> vk::Pipeline {
+        #[cfg(feature = "validation_layers")] {
+            assert!(!self.is_destroyed,
+                    "RenderTargets::pipeline() was called after render_targets destruction");
+        }
         self.pipeline
     }
 
     pub fn swapchain_device(&self) -> &ash::khr::swapchain::Device {
+        #[cfg(feature = "validation_layers")] {
+            assert!(!self.is_destroyed,
+                    "RenderTargets::swapchain_device() was called after render_targets destruction");
+        }
         &self.swapchain_device
     }
 
     pub fn swapchain(&self) -> vk::SwapchainKHR {
+        #[cfg(feature = "validation_layers")] {
+            assert!(!self.is_destroyed,
+                    "RenderTargets::swapchain() was called after render_targets destruction");
+        }
         self.swapchain
     }
 
-    pub unsafe fn destroy(&self, context: &VulkanContext) {
+    pub unsafe fn destroy(&mut self, context: &VulkanContext) {
+        if self.is_destroyed {
+            return;
+        }
+        self.is_destroyed = true;
+
         for framebuffer in self.framebuffers.iter() {
             unsafe { context.device().destroy_framebuffer(*framebuffer, None) };
         }
